Moved utils to services
