+++ components/automate-cli/pkg/verifyserver/services/gcpcloudstorageservice/utils.go
@@ -1,4 +1,4 @@
-package gcputils
+package gcpcloudstorageservice
 
 import (
 	"context"
@@ -16,7 +16,7 @@ type GCPUtils interface {
 	NewSessionWithOptions(ctx context.Context, gsa *models.GcpServiceAccount) (*storage.Client, error)
 	DeleteObject(ctx context.Context, obj *storage.ObjectHandle) error
 	ListObjects(ctx context.Context, bucket *storage.BucketHandle, query *storage.Query) error
-	NewUploader(ctx context.Context, obj *storage.ObjectHandle) (*models.Checks, error)
+	NewUploader(ctx context.Context, uploadObject *storage.ObjectHandle) error
 	BucketAttributes(ctx context.Context, bucket *storage.BucketHandle) error
 }
 
@@ -28,17 +28,21 @@ type GCPUtilsImpl struct {
 }
 
 func (au *GCPUtilsImpl) NewSessionWithOptions(ctx context.Context, gsa *models.GcpServiceAccount) (*storage.Client, error) {
-	bx, _ := json.Marshal(gsa)
-	return storage.NewClient(ctx, option.WithCredentialsJSON(bx))
-}
-
-func (au *GCPUtilsImpl) NewUploader(ctx context.Context, uploadObject *storage.ObjectHandle) (*models.Checks, error) {
-	wc := uploadObject.NewWriter(ctx)
-	_, err := fmt.Fprintf(wc, "Heute ist ein schöner Tag.")
+	bx, err := json.Marshal(gsa)
 	if err != nil {
 		return nil, err
 	}
-	return nil, wc.Close()
+	return storage.NewClient(ctx, option.WithCredentialsJSON(bx))
+}
+
+func (au *GCPUtilsImpl) NewUploader(ctx context.Context, uploadObject *storage.ObjectHandle) error {
+	wc := uploadObject.NewWriter(context.Background())
+	_, err := fmt.Fprintf(wc, "Today is a beautiful day!")
+	if err != nil {
+		return err
+	}
+
+	return nil
 }
 
 func (au *GCPUtilsImpl) DeleteObject(ctx context.Context, obj *storage.ObjectHandle) error {
@@ -52,7 +56,7 @@ func (au *GCPUtilsImpl) ListObjects(ctx context.Context, bucket *storage.BucketH
 			logrus.Errorf("Error listing the objects")
 			return err
 		}
-		fmt.Println("The bucket", attrs.Name)
+		fmt.Println("The object", attrs.Name)
 	}
 	return nil
 }
